Fix encoding issues when the HMAC contains one-or-more null bytes.
I'm not sure why there was so much complexity surrounding passing values
into each function, but it was causing issues whenever the output of the
HMAC, which is binary, contained null bytes.

Null bytes can't be represented by a Bash string, so we would see:
  -bash: warning: command substitution: ignored null byte in input

These would then be ignored, and cause the output to be incorrect.

Now we let each function read directly from stdin, bypassing the need to
first read the input into a string.